Restructured project folders to include Classes, Services, and app-component folders
