+++ frontend/src/app/Classes/user.spec.ts
