+++ frontend/src/app/Classes/user.ts
