+++ frontend/src/app/Services/general.service.ts
