+++ frontend/src/app/Services/user-service.service.spec.ts
@@ -1,13 +1,13 @@
 import { TestBed } from '@angular/core/testing';
 
-import { UserServiceService } from './user-service.service';
+import { UserService } from './user-service.service';
 
 describe('UserServiceService', () => {
-  let service: UserServiceService;
+  let service: UserService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({});
-    service = TestBed.inject(UserServiceService);
+    service = TestBed.inject(UserService);
   });
 
   it('should be created', () => {
